Move BuildEnvironment out of wheel.py
