+++ src/pip/_internal/wheel.py
@@ -17,7 +17,6 @@ import sys
 import warnings
 from base64 import urlsafe_b64encode
 from email.parser import Parser
-from sysconfig import get_paths
 
 from pip._vendor import pkg_resources, pytoml
 from pip._vendor.distlib.scripts import ScriptMaker
@@ -25,6 +24,7 @@ from pip._vendor.packaging.utils import canonicalize_name
 from pip._vendor.six import StringIO
 
 from pip._internal import pep425tags
+from pip._internal.build_env import BuildEnvironment
 from pip._internal.download import path_to_url, unpack_url
 from pip._internal.exceptions import (
     InstallationError, InvalidWheelFilename, UnsupportedWheel,
@@ -600,58 +600,6 @@ class Wheel(object):
         if tags is None:  # for mock
             tags = pep425tags.get_supported()
         return bool(set(tags).intersection(self.file_tags))
-
-
-class BuildEnvironment(object):
-    """Context manager to install build deps in a simple temporary environment
-    """
-    def __init__(self, no_clean):
-        self._temp_dir = TempDirectory(kind="build-env")
-        self._no_clean = no_clean
-
-    def __enter__(self):
-        self._temp_dir.create()
-
-        self.save_path = os.environ.get('PATH', None)
-        self.save_pythonpath = os.environ.get('PYTHONPATH', None)
-
-        install_scheme = 'nt' if (os.name == 'nt') else 'posix_prefix'
-        install_dirs = get_paths(install_scheme, vars={
-            'base': self._temp_dir.path,
-            'platbase': self._temp_dir.path,
-        })
-
-        scripts = install_dirs['scripts']
-        if self.save_path:
-            os.environ['PATH'] = scripts + os.pathsep + self.save_path
-        else:
-            os.environ['PATH'] = scripts + os.pathsep + os.defpath
-
-        if install_dirs['purelib'] == install_dirs['platlib']:
-            lib_dirs = install_dirs['purelib']
-        else:
-            lib_dirs = install_dirs['purelib'] + os.pathsep + \
-                install_dirs['platlib']
-        if self.save_pythonpath:
-            os.environ['PYTHONPATH'] = lib_dirs + os.pathsep + \
-                self.save_pythonpath
-        else:
-            os.environ['PYTHONPATH'] = lib_dirs
-
-        return self._temp_dir.path
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if not self._no_clean:
-            self._temp_dir.cleanup()
-        if self.save_path is None:
-            os.environ.pop('PATH', None)
-        else:
-            os.environ['PATH'] = self.save_path
-
-        if self.save_pythonpath is None:
-            os.environ.pop('PYTHONPATH', None)
-        else:
-            os.environ['PYTHONPATH'] = self.save_pythonpath
 
 
 class WheelBuilder(object):
